Update README.md
📚 Exciting News! Building My Data Portfolio 🚀📊

Today, I am thrilled to announce the creation of my comprehensive data portfolio, showcasing both my expertise and passion for elevating data-driven solutions in the business world. 📈💼

To kickstart this journey, I engaged in an intriguing exploratory data analysis project using Python, delving into the sales data of a bike store for the year 2017. 🚴‍♂️📅 This time, my approach was unique - instead of a traditional statistical analysis, I tackled the data by answering imaginary questions posed by the esteemed CEO of the company. 💭👔

Here are the questions and their insightful answers: